JEXL-127; added support for 'function', includes variable hoisting
git-svn-id: https://svn-us.apache.org/repos/asf/commons/proper/jexl/trunk@1243029 13f79535-47bb-0310-9956-ffa450edef68

+++ src/test/java/org/apache/commons/jexl3/ArrayLiteralTest.java
@@ -51,7 +51,7 @@ public class ArrayLiteralTest extends JexlTestCase {
         JexlContext jc = new MapContext();
 
         Object o = e.evaluate( jc );
-        Object[] check = { new Float(5), new Integer(10) };
+        Object[] check = { new Double(5), new Integer(10) };
         assertTrue( Arrays.equals(check, (Object[])o) );
         assertTrue (o.getClass().isArray() && o.getClass().getComponentType().equals(Number.class));
     }
